Move all state logic to the state namespace
