--- src/Exceptions/StatePersisterMissingValueException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Pinnacle\OpenIdConnect\Exceptions;
-
-class StatePersisterMissingValueException extends OpenIdConnectException
-{
-}
