+++ src/State/Exceptions/StatePersisterMissingValueException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Pinnacle\OpenIdConnect\State\Exceptions;
+
+use Pinnacle\OpenIdConnect\Exceptions\OpenIdConnectException;
+
+class StatePersisterMissingValueException extends OpenIdConnectException
+{
+}
